chore: takeover the project

+++ reset.sh
@@ -5,4 +5,9 @@ if [ "$1" != "" ]; then
     cd "$1"
 fi
 
+if [ -r src ]; then
+    echo "In the main folder, are you foolish ?"
+    exit 1
+fi
+
 rsync -a --delete -i ../../test/data/system_test/ .
